+++ qualtran/bloqs/optimization/max_k_xor_sat/__init__.py
@@ -11,5 +11,5 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from .guiding_state import GuidingState, SimpleGuidingState
+from .kikuchi_guiding_state import GuidingState, SimpleGuidingState
 from .kxor_instance import Constraint, KXorInstance
